+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - new flag `archived` on `project list` to list archived projects
 - a new global config `allow-project-name` that, when enabled, allow the user to the project
   name (or parts of it) to be used where the project id is asked.
+- common function to get all pages on a paginated request, to not reimplement it, and guarantee
+  all entities are being used/returned.
 
 ### Fixed